refactor: split overlay view components
